requirements.txt
pillow>=10.0
numpy>=1.24
qrcode>=7.0
